refactor: Move Project setting context provider to pages.
