+++ src/admin/pages/Project/Context/index.tsx
@@ -1,17 +1,15 @@
 import { ProjectSetting } from '@shared/types';
 import axios from 'axios';
 import React, { createContext, useContext } from 'react';
-import useSWR, { SWRConfiguration, SWRResponse } from 'swr';
+import useSWR, { SWRConfiguration } from 'swr';
 import useSWRMutation, { SWRMutationResponse } from 'swr/mutation';
+import { ProjectSettingContext } from './type';
 
-type ContextType = {
-  getProjectSetting: (config?: SWRConfiguration) => SWRResponse<ProjectSetting>;
-  updateProjectSetting: () => SWRMutationResponse;
-};
+const Context = createContext({} as ProjectSettingContext);
 
-const Context = createContext<ContextType>({} as any);
-
-export const ProjectSettingContextProvider = ({ children }) => {
+export const ProjectSettingContextProvider: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
   const getProjectSetting = (config?: SWRConfiguration) =>
     useSWR(
       '/api/project_settings',
