+++ src/admin/pages/Project/Context/type.ts
@@ -1,0 +1,8 @@
+import { ProjectSetting } from '@shared/types';
+import { SWRConfiguration, SWRResponse } from 'swr';
+import { SWRMutationResponse } from 'swr/mutation';
+
+export type ProjectSettingContext = {
+  getProjectSetting: (config?: SWRConfiguration) => SWRResponse<ProjectSetting>;
+  updateProjectSetting: () => SWRMutationResponse;
+};
